Change how the singleton instance of ThemeSupport is accessed
`WC()->theme_support` can't be used to access the instance of
the ThemeSupport class, since the constructor of the WooCommerce class
ends up executing WC() as well, and this causes a recursion problem.
So the method has been made static, therefore it's
`WooCommerce::theme_support` instead of `WC()->theme_support` now.